Checkpoint 1
This commit contains documentation describing the repository creation process using Gitkraken. The documentation was created using the template created with the commit titled "Creation of Documentation Template". Various screenshots have been used to help illustrate the required actions expected out of the user. The expected audience is that of a user who has previous experience with Github and in fact has an account ready to go.