Checkpoint 4
This commit contains documentation describing the process required to add a user using the GitHub Online Interface.. The documentation was created using the template created with the commit titled "Creation of Documentation Template". Various screenshots have been used to help illustrate the required actions expected out of the user. The user is expected to have already established a repo using alternative methods before using the documentation as the repo creation process is stipulated in the documentation uploaded with the commit "Checkpoint 1".